1-init_dog.c

+++ 0x0E-structures_typedef/1-init_dog.c
@@ -1,4 +1,5 @@
 #include "dog.h"
+#include <stddef.h>
 
 /**
  * init_dog - initializes a dog struct
@@ -9,7 +10,7 @@
  * Return: 0
  */
 
-void init_dog(struct dog *d, char *name, float age, char *owner)
+void init_dog(struct dog *d, char *name, double age, char *owner)
 {
 	if (d != NULL)
 	{
